requirements.txt
alembic==1.14.0
bidict==0.23.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
Flask-SocketIO==5.5.0
Flask-SQLAlchemy==3.1.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0
